axios/sql/requirement.txt
absl-py==1.4.0
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
banal==1.0.6
black==23.3.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
comm==0.1.3
comtypes==1.2.0
contourpy==1.1.0
cycler==0.11.0
dataset==1.6.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
Django==4.2.4
docx2pdf==0.1.8
executing==1.2.0
fastapi==0.100.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-cloud-language==2.10.1
google-pasta==0.2.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.2
grpcio-status==1.56.2
h11==0.14.0
h5py==3.9.0
idna==3.4
ipykernel==6.23.3
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
metrics==0.3.3
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathlib2==2.3.7.post1
pathspec==0.5.5
pdf2image==1.16.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.0.2
pydantic_core==2.1.2
Pygments==2.2.0
PyJWT==2.8.0
pyparsing==3.0.9
pypdf==3.12.1
PyPDF2==3.0.1
pypdfium2==4.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
unicorn==2.0.1.post1
urllib3==1.26.16
uvicorn==0.22.0
validators==0.22.0
virtualenv==20.24.2
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
